+++ tests/when_exist_all_test.go
@@ -1,0 +1,82 @@
+package tests
+
+import (
+	"testing"
+
+	"github.com/golodash/galidator"
+)
+
+func TestWhenExistAll(t *testing.T) {
+	v := g.Validator(g.R().Complex(galidator.Rules{
+		"id":       g.R().Int(),
+		"name":     g.R().String(),
+		"username": g.R().WhenExistAll("id", "name").String().SpecificMessages(galidator.Messages{"when_exist_all": "we are required now"}),
+	}))
+
+	scenarios := []scenario{
+		{
+			name:      "fail",
+			validator: v,
+			in: map[string]interface{}{
+				"id":       1,
+				"name":     "name",
+				"username": "",
+			},
+			panic:    false,
+			expected: map[string][]string{"username": {"we are required now"}},
+		},
+		{
+			name:      "pass-1",
+			validator: v,
+			in: map[string]interface{}{
+				"id":       0,
+				"name":     "",
+				"username": "",
+			},
+			panic:    false,
+			expected: nil,
+		},
+		{
+			name:      "pass-2",
+			validator: v,
+			in: map[string]interface{}{
+				"id":       1,
+				"name":     "",
+				"username": "",
+			},
+			panic:    false,
+			expected: nil,
+		},
+		{
+			name:      "pass-3",
+			validator: v,
+			in: map[string]interface{}{
+				"id":       0,
+				"name":     "name",
+				"username": "",
+			},
+			panic:    false,
+			expected: nil,
+		},
+		{
+			name:      "pass-3",
+			validator: v,
+			in: map[string]interface{}{
+				"id":       1,
+				"name":     "name",
+				"username": "ss",
+			},
+			panic:    false,
+			expected: nil,
+		},
+	}
+
+	for _, s := range scenarios {
+		t.Run(s.name, func(t *testing.T) {
+			defer deferTestCases(t, s.panic, s.expected)
+
+			output := s.validator.Validate(s.in)
+			check(t, s.expected, output)
+		})
+	}
+}